remove outpath if it exists (dont allow for repeated runs)
Careful: It was previously implemented to create a new restrt from the same namelist. However, I did not use this functionality, as the simulation is deterministic given a namelist. I rather run the same namelist repeatedly for debugging pruposes, where I want the old results to be deleted before execution.